Add the feature and refactor

+++ Deduplication.Controller/Extension/GeneralExtension.cs
@@ -4,7 +4,7 @@ namespace Deduplication.Controller.Extensions
 {
     public static class GeneralExtension
     {
-        public static T[] SubArray<T>(this T[] data, int index, int length)
+        public static T[] SubArray<T>(this T[] data, long index, long length)
         {
             T[] result = new T[length];
             Array.Copy(data, index, result, 0, length);
